error handle

+++ routes/route.js
@@ -6,29 +6,46 @@ const mongo = require("mongodb");
 const client = new mongo.MongoClient("mongodb://localhost:27017");
 const db = client.db("shop");
 const items = db.collection("items");
-const connect = async () => {
-    await client.connect()
+const connect = (param) => {
+    try {
+        throw new Error("ble")
+        client.connect();
+    } catch (err) {
+        param(err);
+        return "error";
+    }
 }
 const disconnect = () => {
     client.close();
 }
 
-router.post("/", async (req, res) => {
+router.post("/", async (req, res, next) => {
     const id = Number(req.body);
-    connect();
+    const err = connect(next);
+    if (err === "error") {
+        return;
+    }
     await items.deleteOne({ id });
     await items.updateMany({ id: { $gt: id } }, { $inc: { id: -1 } });
     disconnect();
     res.redirect("back");
 })
 
-router.post("/item/:param", async (req, res) => {
+router.post("/item/:param", async (req, res, next) => {
     const insFun = async (param1, param2, param3, param4) => {               // funkcja zapisująca/modyfikująca przedmiot w bazie
         if (param2 === "create") {
             const form = new multiparty.Form();
             form.parse(req, (err, fields, files) => {
+                if (err) {
+                    next(err);
+                    return;
+                }
                 const ext = files.picture[0].originalFilename.split(".")[1]; // rozszerzenie pliku
                 fs.readFile(files.picture[0].path, async (err, data) => {
+                    if (err) {
+                        next(err);
+                        return;
+                    }
                     const obj = {
                         id: param1,
                         name: fields.name[0],
@@ -37,16 +54,23 @@ router.post("/item/:param", async (req, res) => {
                         price: fields.price[0],
                         desc: fields.desc[0]
                     }
-                    connect();
+                    const err2 = connect(next);
+                    if (err2 === "error") {
+                        return;
+                    }
                     await items.insertOne(obj);
                     disconnect();
+                    res.redirect("back");
                 })
             })
-            res.redirect("back");
         }
         if (param2 === "update") {
             const ext = param4.picture[0].originalFilename.split(".")[1]; // rozszerzenie pliku
             fs.readFile(param4.picture[0].path, async (err, data) => {
+                if (err) {
+                    next(err);
+                    return;
+                }
                 const obj = {
                     id: param1,
                     name: param3.name[0],
@@ -55,7 +79,10 @@ router.post("/item/:param", async (req, res) => {
                     price: param3.price[0],
                     desc: param3.desc[0]
                 }
-                connect();
+                const err2 = connect(next);
+                if (err2 === "error") {
+                    return;
+                }
                 await items.updateOne({ id: param1 }, {
                     $set: obj
                 })
@@ -65,7 +92,10 @@ router.post("/item/:param", async (req, res) => {
         }
     }
     if (req.params.param === "create") {
-        connect();
+        const err = connect(next);
+        if (err === "error") {
+            return;
+        }
         const itemArr = await items.find().toArray();
         disconnect();
         let id;
@@ -78,7 +108,10 @@ router.post("/item/:param", async (req, res) => {
         }
     } else {
         if (req.body.picture === "") {
-            connect();
+            const err = connect(next);
+            if (err === "error") {
+                return;
+            }
             await items.updateOne({ id: Number(req.body.hidden) }, {
                 $set: {
                     name: req.body.name,
@@ -91,6 +124,10 @@ router.post("/item/:param", async (req, res) => {
         } else {
             const form = new multiparty.Form();
             form.parse(req, (err, fields, files) => {
+                if (err) {
+                    next(err);
+                    return;
+                }
                 insFun(Number(fields.hidden[0]), "update", fields, files);
             })
         }
@@ -98,23 +135,32 @@ router.post("/item/:param", async (req, res) => {
 })
 
 router.get("/items", async (req, res, next) => {
-    connect();
+    const err = connect(next);
+    if (err === "error") {
+        return;
+    }
     const itemArr = await items.find().toArray();
     disconnect();
     res.send(itemArr);
 })
 
-router.post("/search", async (req, res) => {
+router.post("/search", async (req, res, next) => {
     const reg = new RegExp(`^${req.body.trim()}`, "i");
-    connect();
+    const err = connect(next);
+    if (err === "error") {
+        return;
+    }
     const itemArr = await items.find({ name: reg }).toArray();
     disconnect();
     res.send(itemArr);
 })
 
-router.post("/all", async (req, res) => {
+router.post("/all", async (req, res, next) => {
     let itemArr;
-    connect();
+    const err = connect(next);
+    if (err === "error") {
+        return;
+    }
     if (req.body === "show") {
         itemArr = await items.find().toArray();
     } else {
@@ -125,8 +171,11 @@ router.post("/all", async (req, res) => {
     res.send(itemArr);
 })
 
-router.post("/update", async (req, res) => {
-    connect();
+router.post("/update", async (req, res, next) => {
+    const err = connect(next);
+    if (err === "error") {
+        return;
+    }
     const itemArr = await items.find({ id: Number(req.body) }).toArray();
     disconnect();
     res.send(itemArr[0]);
